added a length check - console print
checks the length of the text retrieved from the window using the DISCORD_RPC_LIMIT constant, which is set to 128. If the text length exceeds this limit, it is truncated to this limit. If the song name couldn't be read from the window, a message is printed to the console and an empty string is returned.